Move GetClusterID into pkg.k8sapi
